move getSummaryOptions to adapterExtender
